Redecte code

+++ app/posts/crud.py
@@ -3,17 +3,19 @@ from fastapi import Depends, HTTPException, status
 from sqlalchemy import Result, select, update
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from .schemas import CreatePost, UpdatePost, BaseModel
+from typing import List
+
+from .schemas import CreatePost, UpdatePost, BaseModel, PostBase
 from core.models import Post
 
-async def create_post(session:AsyncSession, post_create:CreatePost):
+async def create_post(session:AsyncSession, post_create:CreatePost)->CreatePost:
     post=Post(**post_create.model_dump())
     session.add(post)
     await session.commit()
     await session.refresh(post)
     return post_create
 
-async def delete_all(session:AsyncSession):
+async def delete_all(session:AsyncSession)->None:
     stm = select(Post).order_by(Post.id)
     result :Result  =await session.execute(stm)
     posts = result.scalars().all()
@@ -21,12 +23,16 @@ async def delete_all(session:AsyncSession):
         await session.delete(post)
     await session.commit()
 
-async def delete_by_id(session:AsyncSession, post_id:int, user_id:int):
-    post = await session.get(Post, post_id)
+async def delete_by_id(session:AsyncSession, post:Post, user_id:int)->None:
+    if post.user_id!=user_id:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Is not your post"
+        )
     await session.delete(post)
     await session.commit()
 
-async def delete_by_user_id(session:AsyncSession, user_id:int):
+async def delete_by_user_id(session:AsyncSession, user_id:int)->None:
     stm = select(Post).where(Post.user_id==user_id)
     result :Result  =await session.execute(stm)
     posts = result.scalars().all()
@@ -34,7 +40,7 @@ async def delete_by_user_id(session:AsyncSession, user_id:int):
         await session.delete(post)
     await session.commit()
 
-async def get_all(session:AsyncSession):
+async def get_all(session:AsyncSession)->List[PostBase]:
     stm = select(Post).order_by(Post.id)
     result :Result  =await session.execute(stm)
     return result.scalars().all()
